:recycle: move `maybe_coroutine` to private
