+++ discord/commands/core.py
@@ -68,8 +68,8 @@ from ..object import Object
 from ..role import Role
 from ..threads import Thread
 from ..user import User
-from ..utils import MISSING, find, maybe_coroutine, utcnow
-from ..utils.private import warn_deprecated, async_all
+from ..utils import MISSING, find, utcnow
+from ..utils.private import warn_deprecated, async_all, maybe_coroutine
 from .context import ApplicationContext, AutocompleteContext
 from .options import Option, OptionChoice
