+++ discord/ext/commands/flags.py
@@ -32,8 +32,8 @@ from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any, Iterator, Literal, Pattern, TypeVar, Union
 
 from discord import utils
-from ...utils import MISSING, Undefined, maybe_coroutine
-from ...utils.private import resolve_annotation
+from ...utils import MISSING, Undefined
+from ...utils.private import resolve_annotation, maybe_coroutine
 
 from .converter import run_converters
 from .errors import (
